snapshot_manager.py
"""
FASE 1: Snapshot Manager para Qdrant
Proporciona capacidades de backup automático con compresión y restauración.

Funcionalidades:
- Crear snapshots bajo demanda
- Snapshots automáticos programados (diarios)
- Compresión de snapshots (GZIP)
- Listado y gestión de snapshots
- Restauración desde snapshot
- Validación de integridad
"""

import os
import json
import gzip
import shutil
from datetime import datetime, timedelta
from pathlib import Path
from typing import Optional, Dict, List
from .vector_store import get_or_create_vector_store
from dotenv import load_dotenv

load_dotenv()


class SnapshotManager:
    """
    Gestor de snapshots para Qdrant con compresión y validación.
    
    Almacena snapshots en directorio configurable con estructura:
    snapshots/
    ├── voia_vectors/
    │   ├── snapshot_2025-11-13_14-30-00.tar.gz
    │   ├── snapshot_2025-11-13_14-30-00.manifest
    │   └── ...
    └── metadata.json (índice global)
    """

    def __init__(
        self,
        snapshots_dir: Optional[str] = None,
        collection_name: str = "voia_vectors",
        compress: bool = True,
        retention_days: int = 30
    ):
        """
        Inicializa el SnapshotManager.
        
        Args:
            snapshots_dir: Directorio para almacenar snapshots (default: ./snapshots)
            collection_name: Nombre de la colección en Qdrant (default: voia_vectors)
            compress: Comprimir snapshots con GZIP (default: True)
            retention_days: Días de retención para snapshots antiguos (default: 30)
        """
        self.client = get_or_create_vector_store()
        self.collection_name = collection_name
        self.compress = compress
        self.retention_days = retention_days

        # Configurar directorio de snapshots
        self.snapshots_dir = Path(
            snapshots_dir or os.getenv("QDRANT_SNAPSHOTS_DIR", "./snapshots")
        )
        self.snapshots_dir.mkdir(parents=True, exist_ok=True)

        # Subdirectorio para cada colección
        self.collection_dir = self.snapshots_dir / collection_name
        self.collection_dir.mkdir(parents=True, exist_ok=True)

        # Archivo de metadatos global
        self.metadata_file = self.snapshots_dir / "metadata.json"
        self._init_metadata()

    def _init_metadata(self) -> None:
        """Inicializa o carga archivo de metadatos."""
        if not self.metadata_file.exists():
            self.metadata = {
                "created": datetime.now().isoformat(),
                "snapshots": {}
            }
            self._save_metadata()
        else:
            try:
                with open(self.metadata_file, 'r') as f:
                    self.metadata = json.load(f)
            except Exception as e:
                print(f"⚠️ Error cargando metadatos: {e}, inicializando nuevo")
                self.metadata = {
                    "created": datetime.now().isoformat(),
                    "snapshots": {}
                }

    def _save_metadata(self) -> None:
        """Guarda metadatos en archivo."""
        try:
            with open(self.metadata_file, 'w') as f:
                json.dump(self.metadata, f, indent=2, default=str)
        except Exception as e:
            print(f"❌ Error guardando metadatos: {e}")

    def create_snapshot(self, description: str = "") -> Dict:
        """
        Crea un snapshot de la colección Qdrant.
        
        ⚠️ NOTA: Los snapshots no están soportados en esta versión.
        """
        print(f"⚠️  Snapshots no disponibles - skipping")
        return {
            "success": False,
            "error": "Snapshots not available in this Qdrant version",
            "collection": self.collection_name,
            "timestamp": datetime.now().isoformat()
        }

    def restore_snapshot(self, snapshot_id: str) -> Dict:
        """
        Restaura una colección desde un snapshot.
        
        ⚠️ Snapshots no están disponibles.
        """
        return {
            "success": False,
            "error": "Snapshots not available",
        }

    def list_snapshots(self, collection: Optional[str] = None) -> List[Dict]:
        """
        Lista todos los snapshots disponibles.
        """
        return []

    def _cleanup_old_snapshots(self) -> None:
        """
        Elimina snapshots más antiguos que retention_days.
        """
        pass

    def validate_snapshot(self, snapshot_id: str) -> Dict:
        """
        Valida integridad de un snapshot.
        """
        return {
            "valid": False,
            "error": "Snapshots not available",
        }

    def get_statistics(self) -> Dict:
        """Obtiene estadísticas de snapshots."""
        return {
            "total_snapshots": 0,
            "total_size_mb": 0,
            "retention_days": self.retention_days,
            "compress_enabled": self.compress,
        }

