requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
sqlalchemy
mysql-connector-python
pydantic
requests
qdrant-client
sentence_transformers
beautifulsoup4
slowapi
redis